Managing redux data in the Container component
