+++ src/components/Dashboard.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import SideBar from './SideBar';
+
+const Dashboard = ({ company }) => (
+  <div>
+    <SideBar />
+  </div>
+);
+
+export default Dashboard;
