+++ src/components/SideBar.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const SideBar = () => (
+  <div>
+    SideBar
+  </div>
+);
+
+export default SideBar;
